test(plugin): mocks have a pre-existing shared directory
